+++ cloudos/__init__.py
@@ -6,5 +6,6 @@ Python package for interacting with Cloud OS (https://cloudos.lifebit.ai/)
 """
 
 from .clos import Cloudos
+from ._version import __version__
 
-__all__ = ['jobs', 'utils', 'clos']
+__all__ = ['jobs', 'utils', 'clos', 'cohorts']