+++ cloudos/__main__.py
@@ -6,24 +6,39 @@ from cloudos.clos import Cloudos
 import json
 import time
 import sys
-
+from ._version import __version__
 
 # GLOBAL VARS
 JOB_COMPLETED = 'completed'
 JOB_FAILED = 'failed'
 JOB_ABORTED = 'aborted'
+REQUEST_INTERVAL = 30
 
 
 @click.group()
+@click.version_option(__version__)
 def run_cloudos_cli():
     """CloudOS python package: a package for interacting with CloudOS."""
     print(run_cloudos_cli.__doc__ + '\n')
+    print('Version: ' + __version__ + '\n')
 
 
 @run_cloudos_cli.group()
 def job():
     """CloudOS job functionality: run and check jobs in CloudOS."""
     print(job.__doc__ + '\n')
+
+
+@run_cloudos_cli.group()
+def workflow():
+    """CloudOS workflow functionality: list workflows in CloudOS."""
+    print(workflow.__doc__ + '\n')
+
+
+@run_cloudos_cli.group()
+def cromwell():
+    """Cromwell server functionality: check status, start and stop."""
+    print(cromwell.__doc__ + '\n')
 
 
 @job.command('run')
@@ -80,6 +95,15 @@ def job():
 @click.option('--spot',
               help='Whether to make a spot instance.',
               is_flag=True)
+@click.option('--storage-mode',
+              help=('Either \'lustre\' or \'regular\'. Indicates if the user wants to select ' +
+                    'regular or lustre storage. Default=regular.'),
+              default='regular')
+@click.option('--lustre-size',
+              help=('The lustre storage to be used when --storage-mode=lustre, in GB. It should ' +
+                    'be 1200 or a multiple of it. Default=1200.'),
+              type=int,
+              default=1200)
 @click.option('--wait-completion',
               help=('Whether to wait to job completion and report final ' +
                     'job status.'),
@@ -106,6 +130,8 @@ def run(apikey,
         instance_type,
         instance_disk,
         spot,
+        storage_mode,
+        lustre_size,
         wait_completion,
         wait_time,
         verbose):
@@ -113,7 +139,7 @@ def run(apikey,
     print('Executing run...')
     if verbose:
         print('\t...Preparing objects')
-    j = jb.Job(apikey, cloudos_url, workspace_id, project_name, workflow_name)
+    j = jb.Job(cloudos_url, apikey, None, workspace_id, project_name, workflow_name)
     if verbose:
         print('\tThe following Job object was created:')
         print('\t' + str(j))
@@ -127,7 +153,9 @@ def run(apikey,
                       nextflow_profile,
                       instance_type,
                       instance_disk,
-                      spot)
+                      spot,
+                      storage_mode,
+                      lustre_size)
     print(f'\tYour assigned job id is: {j_id}')
     j_url = f'{cloudos_url}/app/jobs/{j_id}'
     if wait_completion:
@@ -149,11 +177,11 @@ def run(apikey,
                 print(f'\tYour job took {elapsed} seconds to abort.')
                 sys.exit(1)
             else:
-                elapsed += 1
+                elapsed += REQUEST_INTERVAL
                 if j_status_h != j_status_h_old:
                     print(f'\tYour current job status is: {j_status_h}.')
                     j_status_h_old = j_status_h
-                time.sleep(60)
+                time.sleep(REQUEST_INTERVAL)
         j_status = j.get_job_status(j_id)
         j_status_h = json.loads(j_status.content)["status"]
         if j_status_h != JOB_COMPLETED:
@@ -195,15 +223,15 @@ def run(apikey,
 @click.option('--verbose',
               help='Whether to print information messages or not.',
               is_flag=True)
-def status(apikey,
-           cloudos_url,
-           job_id,
-           verbose):
+def job_status(apikey,
+               cloudos_url,
+               job_id,
+               verbose):
     """Check job status in CloudOS."""
     print('Executing status...')
     if verbose:
         print('\t...Preparing objects')
-    cl = Cloudos(apikey, cloudos_url)
+    cl = Cloudos(cloudos_url, apikey, None)
     if verbose:
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
@@ -229,39 +257,249 @@ def status(apikey,
 @click.option('--workspace-id',
               help='The specific CloudOS workspace id.',
               required=True)
-@click.option('--outfile',
-              help=('Output filename where to save job table. ' +
-                    'Default=joblist.csv'),
-              default='joblist.csv',
+@click.option('--output-basename',
+              help=('Output file base name to save jobs list. ' +
+                    'Default=joblist'),
+              default='joblist',
               required=False)
-@click.option('--full-data',
-              help=('Whether to collect full available data from jobs or ' +
-                    'just the preconfigured selected fields.'),
+@click.option('--output-format',
+              help='The desired file format (file extension) for the output.',
+              type=click.Choice(['csv', 'json'], case_sensitive=False),
+              default='csv')
+@click.option('--all-fields',
+              help=('Whether to collect all available fields from jobs or ' +
+                    'just the preconfigured selected fields. Only applicable ' +
+                    'when --output-format=csv'),
               is_flag=True)
 @click.option('--verbose',
               help='Whether to print information messages or not.',
               is_flag=True)
-def list(apikey,
-         cloudos_url,
-         workspace_id,
-         outfile,
-         full_data,
-         verbose):
+def list_jobs(apikey,
+              cloudos_url,
+              workspace_id,
+              output_basename,
+              output_format,
+              all_fields,
+              verbose):
     """Collect all your jobs from a CloudOS workspace in CSV format."""
+    outfile = output_basename + '.' + output_format
     print('Executing list...')
     if verbose:
         print('\t...Preparing objects')
-    cl = Cloudos(apikey, cloudos_url)
+    cl = Cloudos(cloudos_url, apikey, None)
     if verbose:
         print('\tThe following Cloudos object was created:')
         print('\t' + str(cl) + '\n')
         print('\tSearching for jobs in the following workspace: ' +
               f'{workspace_id}')
     my_jobs_r = cl.get_job_list(workspace_id)
-    my_jobs = cl.process_job_list(my_jobs_r, full_data)
-    my_jobs.to_csv(outfile, index=False)
-    print(f'\tJob list collected with a total of {my_jobs.shape[0]} jobs.')
-    print(f'\tJob list table saved to {outfile}')
+    if output_format == 'csv':
+        my_jobs = cl.process_job_list(my_jobs_r, all_fields)
+        my_jobs.to_csv(outfile, index=False)
+        print(f'\tJob list collected with a total of {my_jobs.shape[0]} jobs.')
+    elif output_format == 'json':
+        with open(outfile, 'w') as o:
+            o.write(my_jobs_r.text)
+        print(f'\tJob list collected with a total of {len(json.loads(my_jobs_r.content)["jobs"])} jobs.')
+    else:
+        raise ValueError('Unrecognised output format. Please use one of [csv|json]')
+    print(f'\tJob list saved to {outfile}')
+
+
+@workflow.command('list')
+@click.option('-k',
+              '--apikey',
+              help='Your CloudOS API key',
+              required=True)
+@click.option('-c',
+              '--cloudos-url',
+              help=('The CloudOS url you are trying to access to. ' +
+                    'Default=https://cloudos.lifebit.ai.'),
+              default='https://cloudos.lifebit.ai')
+@click.option('--workspace-id',
+              help='The specific CloudOS workspace id.',
+              required=True)
+@click.option('--output-basename',
+              help=('Output file base name to save workflow list. ' +
+                    'Default=workflow_list'),
+              default='workflow_list',
+              required=False)
+@click.option('--output-format',
+              help='The desired file format (file extension) for the output.',
+              type=click.Choice(['csv', 'json'], case_sensitive=False),
+              default='csv')
+@click.option('--all-fields',
+              help=('Whether to collect all available fields from workflows or ' +
+                    'just the preconfigured selected fields. Only applicable ' +
+                    'when --output-format=csv'),
+              is_flag=True)
+@click.option('--verbose',
+              help='Whether to print information messages or not.',
+              is_flag=True)
+def list_workflows(apikey,
+                   cloudos_url,
+                   workspace_id,
+                   output_basename,
+                   output_format,
+                   all_fields,
+                   verbose):
+    """Collect all workflows from a CloudOS workspace in CSV format."""
+    outfile = output_basename + '.' + output_format
+    print('Executing list...')
+    if verbose:
+        print('\t...Preparing objects')
+    cl = Cloudos(cloudos_url, apikey, None)
+    if verbose:
+        print('\tThe following Cloudos object was created:')
+        print('\t' + str(cl) + '\n')
+        print('\tSearching for workflows in the following workspace: ' +
+              f'{workspace_id}')
+    my_workflows_r = cl.get_workflow_list(workspace_id)
+    if output_format == 'csv':
+        my_workflows = cl.process_workflow_list(my_workflows_r, all_fields)
+        my_workflows.to_csv(outfile, index=False)
+        print(f'\tWorkflow list collected with a total of {my_workflows.shape[0]} workflows.')
+    elif output_format == 'json':
+        with open(outfile, 'w') as o:
+            o.write(my_workflows_r.text)
+        print(f'\tWorkflow list collected with a total of {len(json.loads(my_workflows_r.content))} workflows.')
+    else:
+        raise ValueError('Unrecognised output format. Please use one of [csv|json]')
+    print(f'\tWorkflow list saved to {outfile}')
+
+
+@cromwell.command('status')
+@click.version_option()
+@click.option('-t',
+              '--cromwell-token',
+              help='Specific Cromwell server authentication token.',
+              required=True)
+@click.option('-c',
+              '--cloudos-url',
+              help=('The CloudOS url you are trying to access to. ' +
+                    'Default=https://cloudos.lifebit.ai.'),
+              default='https://cloudos.lifebit.ai')
+@click.option('--workspace-id',
+              help='The specific CloudOS workspace id.',
+              required=True)
+@click.option('--verbose',
+              help='Whether to print information messages or not.',
+              is_flag=True)
+def cromwell_status(cromwell_token,
+                    cloudos_url,
+                    workspace_id,
+                    verbose):
+    """Check Cromwell server status in CloudOS."""
+    print('Executing status...')
+    if verbose:
+        print('\t...Preparing objects')
+    cl = Cloudos(cloudos_url, None, cromwell_token)
+    if verbose:
+        print('\tThe following Cloudos object was created:')
+        print('\t' + str(cl) + '\n')
+        print(f'\tChecking Cromwell status in {workspace_id} workspace')
+    c_status = cl.get_cromwell_status(workspace_id)
+    c_status_h = json.loads(c_status.content)["status"]
+    print(f'\tCurrent Cromwell server status is: {c_status_h}\n')
+
+
+@cromwell.command('start')
+@click.version_option()
+@click.option('-t',
+              '--cromwell-token',
+              help='Specific Cromwell server authentication token.',
+              required=True)
+@click.option('-c',
+              '--cloudos-url',
+              help=('The CloudOS url you are trying to access to. ' +
+                    'Default=https://cloudos.lifebit.ai.'),
+              default='https://cloudos.lifebit.ai')
+@click.option('--workspace-id',
+              help='The specific CloudOS workspace id.',
+              required=True)
+@click.option('--wait-time',
+              help=('Max time to wait (in seconds) to Cromwell restart. ' +
+                    'Default=300.'),
+              default=300)
+@click.option('--verbose',
+              help='Whether to print information messages or not.',
+              is_flag=True)
+def cromwell_restart(cromwell_token,
+                     cloudos_url,
+                     workspace_id,
+                     wait_time,
+                     verbose):
+    """Restart Cromwell server in CloudOS."""
+    action = 'restart'
+    print('Starting Cromwell server...')
+    if verbose:
+        print('\t...Preparing objects')
+    cl = Cloudos(cloudos_url, None, cromwell_token)
+    if verbose:
+        print('\tThe following Cloudos object was created:')
+        print('\t' + str(cl) + '\n')
+        print(f'\tStarting Cromwell server in {workspace_id} workspace')
+    cl.cromwell_switch(workspace_id, action)
+    c_status = cl.get_cromwell_status(workspace_id)
+    c_status_h = json.loads(c_status.content)["status"]
+    print(f'\tCurrent Cromwell server status is: {c_status_h}\n')
+    elapsed = 0
+    while elapsed < wait_time and (c_status_h == 'Initializing' or c_status_h == 'Setup'):
+        c_status_old = c_status_h
+        time.sleep(REQUEST_INTERVAL)
+        elapsed += REQUEST_INTERVAL
+        c_status = cl.get_cromwell_status(workspace_id)
+        c_status_h = json.loads(c_status.content)["status"]
+        if c_status_h != c_status_old:
+            print(f'\tCurrent Cromwell server status is: {c_status_h}\n')
+    if c_status_h != 'Running':
+        print(f'\tYour current Cromwell status is: {c_status_h}. The ' +
+              f'selected wait-time of {wait_time} was exceeded. Please, ' +
+              'consider to set a longer wait-time.')
+        print('\tTo further check your Cromwell status you can either go to ' +
+              f'{cloudos_url} or use the following command:\n' +
+              'cloudos cromwell status \\\n' +
+              f'    -c {cloudos_url} \\\n' +
+              '    -t $CROMWELL_TOKEN \\\n' +
+              f'    --workspace-id {workspace_id}')
+        sys.exit(1)
+
+
+@cromwell.command('stop')
+@click.version_option()
+@click.option('-t',
+              '--cromwell-token',
+              help='Specific Cromwell server authentication token.',
+              required=True)
+@click.option('-c',
+              '--cloudos-url',
+              help=('The CloudOS url you are trying to access to. ' +
+                    'Default=https://cloudos.lifebit.ai.'),
+              default='https://cloudos.lifebit.ai')
+@click.option('--workspace-id',
+              help='The specific CloudOS workspace id.',
+              required=True)
+@click.option('--verbose',
+              help='Whether to print information messages or not.',
+              is_flag=True)
+def cromwell_stop(cromwell_token,
+                  cloudos_url,
+                  workspace_id,
+                  verbose):
+    """Stop Cromwell server in CloudOS."""
+    action = 'stop'
+    print('Stopping Cromwell server...')
+    if verbose:
+        print('\t...Preparing objects')
+    cl = Cloudos(cloudos_url, None, cromwell_token)
+    if verbose:
+        print('\tThe following Cloudos object was created:')
+        print('\t' + str(cl) + '\n')
+        print(f'\tRestarting Cromwell server in {workspace_id} workspace')
+    cl.cromwell_switch(workspace_id, action)
+    c_status = cl.get_cromwell_status(workspace_id)
+    c_status_h = json.loads(c_status.content)["status"]
+    print(f'\tCurrent Cromwell server status is: {c_status_h}\n')
 
 
 if __name__ == "__main__":