+++ cloudos/jobs/job.py
@@ -101,7 +101,7 @@ class Job(Cloudos):
         allowed_resources = ['projects', 'workflows']
         if resource not in allowed_resources:
             raise ValueError('Your specified resource is not supported. ' +
-                             'Use one of the following: {allowed_resources}')
+                             f'Use one of the following: {allowed_resources}')
         data = {"apikey": apikey}
         r = requests.get("{}/api/v1/{}?teamId={}".format(cloudos_url,
                                                          resource,
@@ -125,7 +125,9 @@ class Job(Cloudos):
                                  nextflow_profile,
                                  instance_type,
                                  instance_disk,
-                                 spot):
+                                 spot,
+                                 storage_mode,
+                                 lustre_size):
         """Converts a nextflow.config file into a json formatted dict.
 
         Parameters
@@ -157,6 +159,12 @@ class Job(Cloudos):
             The disk space of the instance, in GB.
         spot : bool
             Whether to create a spot instance or not.
+        storage_mode : string
+            Either 'lustre' or 'regular'. Indicates if the user wants to select regular
+            or lustre storage.
+        lustre_size : int
+            The lustre storage to be used when --storage-mode=lustre, in GB. It should be 1200 or
+            a multiple of it.
 
         Returns
         -------
@@ -221,6 +229,15 @@ class Job(Cloudos):
                              }
         else:
             revision_block = ""
+        if storage_mode == "lustre":
+            print('\n[WARNING] Lustre storage has been selected. Please, be sure that this kind of ' +
+                  'storage is available in your CloudOS workspace.\n')
+            if lustre_size % 1200:
+                raise ValueError('Please, specify a lustre storage size of 1200 or a multiple of it. ' +
+                                 f'{lustre_size} is not a valid number.')
+        if storage_mode not in ['lustre', 'regular']:
+            raise ValueError('Please, use either \'lustre\' or \'regular\' for --storage-mode ' +
+                             f'{storage_mode} is not allowed')
 
         params = {
             "parameters": workflow_params,
@@ -237,6 +254,8 @@ class Job(Cloudos):
                 "computeCostLimit": -1,
                 "optim": "test"
             },
+            "lusterFsxStorageSizeInGb": lustre_size,
+            "storageMode": storage_mode,
             "revision": revision_block,
             "profile": nextflow_profile,
             instance: instance_type_block,
@@ -250,16 +269,18 @@ class Job(Cloudos):
         return params
 
     def send_job(self,
-                 job_config,
-                 git_commit,
-                 git_tag,
-                 job_name,
-                 resumable,
-                 batch,
-                 nextflow_profile,
-                 instance_type,
-                 instance_disk,
-                 spot):
+                 job_config=None,
+                 git_commit=None,
+                 git_tag=None,
+                 job_name='new_job',
+                 resumable=False,
+                 batch=False,
+                 nextflow_profile=None,
+                 instance_type='c5.xlarge',
+                 instance_disk=500,
+                 spot=False,
+                 storage_mode='regular',
+                 lustre_size=1200):
         """Send a job to CloudOS.
 
         Parameters
@@ -287,6 +308,12 @@ class Job(Cloudos):
             The disk space of the instance, in GB.
         spot : bool
             Whether to create a spot instance or not.
+        storage_mode : string
+            Either 'lustre' or 'regular'. Indicates if the user wants to select regular
+            or lustre storage.
+        lustre_size : int
+            The lustre storage to be used when --storage-mode=lustre, in GB. It should be 1200 or
+            a multiple of it.
 
         Returns
         -------
@@ -314,7 +341,9 @@ class Job(Cloudos):
                                                nextflow_profile,
                                                instance_type,
                                                instance_disk,
-                                               spot)
+                                               spot,
+                                               storage_mode,
+                                               lustre_size)
         r = requests.post("{}/api/v1/jobs?teamId={}".format(cloudos_url,
                                                             workspace_id),
                           data=json.dumps(params), headers=headers)